플레이어 공격 콤보 구현 및 데미지 텍스트 추가
플레이어 공격 콤보 애니메이션 변경
데메지 텍스트 스크립트 추가
데미지 텍스트는 풀링 사용

+++ Assets/Scripts/UI/Presenter/HudPresenter.cs
@@ -56,5 +56,5 @@ public class HUDPresenter : BasePresenter
         => _view.SetSg(cur / max);
 
     private void OnExpChanged(int cur, int max)
-        => _view.SetExp(cur/max);
+        => _view.SetExp((float)cur / max);
 }
